[FEAT/HM-101] 라우트 변경
