+++ src/pages/ResultPage.tsx
@@ -1,26 +1,22 @@
-import { useState } from 'react';
-import { useQuery } from '@tanstack/react-query';
 import Button from '@/components/common/Button';
 import Header from '@/components/common/Header';
+import ResultInfoComp, {
+  BackLayout,
+  Container,
+} from '@/components/meeting/result/ResultInfoComp';
+import ResultNavbar from '@/components/meeting/result/ResultNavbar';
 import ResultTimeTable from '@/components/meeting/result/ResultTimeTable';
-import { ResultHeatmapProps, TimeTableData } from '@/types/timeTableTypes';
+import UrlShareModal from '@/components/meeting/result/UrlShareModal';
+import useModal from '@/hooks/useModal';
 import {
   ButtonContainer,
   NormalContainer,
 } from '@/styles/components/container';
-import { useTimeStore } from '@/store/meeting/useTimeStore';
-import {
-  formatPostDateTime,
-  formatTimeTableData,
-} from '@/utils/meeting/timetable/formatDateTime';
-import AttendStatusHeader from '@/components/meeting/result/AttendStatusHeader';
-import useModal from '@/hooks/useModal';
-import ResultTimeSelectModal from '@/components/meeting/result/ResultTimeSelectModal';
+import { ResultHeatmapProps, TimeTableData } from '@/types/timeTableTypes';
+import { useQuery } from '@tanstack/react-query';
 
 function ResultPage() {
-  const input = ['2024-06-29T10:00:00', '2024-07-07T21:00:00'];
-  const result = formatTimeTableData(input);
-  console.log('result: ', result);
+  const { isOpen, closeModal, openModal } = useModal();
 
   const timeTableData: TimeTableData = {
     hours: [
@@ -53,10 +49,6 @@ function ResultPage() {
     isStartHalfMinute: false,
   };
 
-  const [isSelected, setIsSelected] = useState(false);
-  const { isOpen, closeModal, openModal } = useModal();
-  const { selectedResult } = useTimeStore();
-
   const { isPending, error, data } = useQuery<ResultHeatmapProps>({
     queryKey: ['selectedTimeData'],
     queryFn: () => fetch('/selectedResult').then((res) => res.json()),
@@ -66,46 +58,27 @@ function ResultPage() {
   if (error) return <div>에러가 발생했습니다</div>;
   if (!data) return <div>데이터가 없습니다</div>;
 
-  const handleDecide = () => {
+  const handleClick = () => {
     openModal();
-    setIsSelected(true);
-    console.log('selectedResult: ', formatPostDateTime(selectedResult));
+    console.log('click');
   };
 
   return (
     <NormalContainer>
       <Header title="일정 조율" />
-      <AttendStatusHeader
-        TotalParticipants={data.totalParticipants.names.length}
-        currentParticipants={data.participatedUsers.names.length}
-        participatedUsers={data.participatedUsers.names}
-        unParticipatedUsers={data.totalParticipants.names.filter(
-          (name) => !data.participatedUsers.names.includes(name)
-        )}
-      />
-      <ResultTimeTable
-        data={timeTableData}
-        roomInfo={data}
-        dragDisabled={isSelected}
-      />
+      <ResultNavbar />
+      <ResultInfoComp />
+      <BackLayout>
+        <Container>
+          <ResultTimeTable roomInfo={data} data={timeTableData} dragDisabled />
+        </Container>
+      </BackLayout>
       <ButtonContainer>
-        <Button
-          $style="solid"
-          onClick={handleDecide}
-          $theme="primary"
-          disabled={selectedResult.length === 0}
-        >
-          {selectedResult.length === 0
-            ? '드래그로 시간 확정하기'
-            : '일정 확정하기'}
+        <Button $style="solid" onClick={handleClick}>
+          공유하기
         </Button>
       </ButtonContainer>
-      {isOpen ? (
-        <ResultTimeSelectModal
-          handleModalClose={closeModal}
-          decidedTime={selectedResult}
-        />
-      ) : null}
+      {isOpen && <UrlShareModal handleModalClose={closeModal} />}
     </NormalContainer>
   );
 }
