+++ src/routes/index.tsx
@@ -5,12 +5,11 @@ import NewMeetingPage from '@/pages/NewMeetingPage';
 import MeetingListPage from '@/pages/MeetingListPage';
 import MeetingPage from '@/pages/MeetingPage';
 import SelectPage from '@/pages/SelectPage';
-import ResultPage from '@/pages/ResultPage';
 import MakingRoomPage from '@/pages/MakingRoomPage';
 import ConfirmMeeting from '@/pages/ConfirmMeeting';
-import ResultGraphPage from '@/pages/ResultGraphPage';
 import HomePage from '@/pages/HomePage';
-
+import DecisionPage from '@/pages/DecisionPage';
+import ResultPage from '@/pages/ResultPage';
 
 const router = createBrowserRouter([
   {
@@ -50,12 +49,12 @@ const router = createBrowserRouter([
     element: <SelectPage />,
   },
   {
-    path: 'meeting/:id/result',
-    element: <ResultPage />,
+    path: 'meeting/:id/decision',
+    element: <DecisionPage />,
   },
   {
-    path: 'meeting/:id/result/graph',
-    element: <ResultGraphPage />,
+    path: 'meeting/:id/result',
+    element: <ResultPage />,
   },
 ]);
